Title: A/B 출력, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250912/A/B 출력/a-divide-b.cpp
@@ -7,7 +7,7 @@ int main() {
     int a, b;
     cin >> a >> b;
 
-    double num;
+    int num;
     cout << a/b << ".";
 
     for(int i=0;i<20;i++)
